Document Freestyle Sinus Displacement Phase Cycle
The manual currently does not explain at which value the Phase of the Freestyle Sinus Displacement Geometry Modifier, cycles/repeats. I trial-and-errorred this and found that it is apparently linked to the wavelength and not to a 0-1, 0-360deg, 0-[n]rad or such range.

Given the manual page is pretty empty (so there's room to fill) and the phase behavior seems kind of vital to me in a sinus-based parametric modifier I am now proposing this information be included.

Thanks a lot for the awesome manual and perfectly accurate build instructions etc., very pleasant to contribute. :)

Patch by @simonrepp

Differential Revision: https://developer.blender.org/D5971