Update the addons/sequencer manual pages
This patch updates the power sequencer add-on description, and as it's the only add-on in the sequencer category, I removed the "Work in Progress" mention.

The add-on's full documentation is on our website, so I linked to it. Is it okay if we keep it that way? Note that the manual, like the addon, is free and open source.

Part of the reason why I'd like to keep it that way is that we have a features reference that's generated and updated directly from the python code. It uses go templates to generate the final html page: gdquest.com/docs/power-sequencer/reference/

The other reason is that it's much easier for my teammates and I to update, as we work on that section of the website as part of multiple free software projects.

Author: @gdquest

Differential Revisison: https://developer.blender.org/D6857