Revert minimum sphinx version to 3.3
Sphinx 3.3 is the requirement because we need a
feature from this version to build locales.

Setting this to stricter versions can be frustrating
for writers when there setup suddenly fails.
So, to keep things simple lets keep this requirement as loose as possible.

This is also helpful for writers who use system
packaged versions of sphinx which are updated less often.