Clean up Knife Project documentation
The knife project documentation had not been updated to reflect the August/September 2021 updates that allow multi-object cutting.

In addition, the description of how to use the tool was not clear, as evidenced by several forum posts, YouTube videos, and questions on blender.chat's support channel about the tool.

This diff tries to make the usage of the tool clearer while showing how multiple cutters and objects are used.

Differential Revision: https://developer.blender.org/D13469