Improve "3D Viewport" section (Part 2 - navigation)
Extended/updated a bunch of text for conciseness and clarity - as well as completeness,
because shortcuts like Alt-MMB for setting the viewport's point of interest and X/Z for pitch/roll
correction in Fly Navigation aren't currently documented. Also removed a few old references to layers,
which have been replaced by collections for a while now.

Author: @arc_

Differential Revision: https://developer.blender.org/D15503