Fix incorrect "Snap To Face Project"/"Snap To Face Nearest" documentation
The documentation currently states that "Face Project" can snap to both
visible and occluded geometry, and "Face Nearest" only to visible geometry.
In reality, it's the opposite. (See https://developer.blender.org/rB011327224ece)

Also updated the screenshot, and tweaked the explanations for the other options
a bit (again) - especially the "Volume" option, which I always assumed was for Volume objects.

Author: @arc_

Differential Revision: https://developer.blender.org/D16301