blender_help_extract: rewrite to use Blender's --help output
Use Blender's help message directly instead of parsing the C source code
which was quite involved and fragile. Note that this uses
`bpy.app.help_text(all=True)` which shows all options, even those
not available on the current build. This was a limitation that prevented
the output from `--help` being used in the past.

Uses half the lines of code and reduces complexity considerably.