Add Catmull Rom as supported types to relevant node pages
Add Catmull Rom as supported types to relevant node pages.

Also add some additional information.

Fixes #104434

---

I think the issue here is that the addition of Catmull Rom is not documented. I assume that this came from the fact that Catmull Rom type is only available in the new curves data. I imagine it's tricky to document it (adding it as one of the types [here](https://docs.blender.org/manual/en/4.0/modeling/curves/structure.html) for example) since it's not available in the current (legacy) curve edit mode, etc.

So I only did minimal edit to reduce confusion like #104434. Even if this will only last for 3.6, I think it's worth it because this is an LTS version.

**Spline Resolution**
- Add Catmull Rom as one of the supported types
- According to this commit https://projects.blender.org/blender/blender/commit/ddf956fb5846da3835a1785978e0b9f1af89b45d, poly type splines can have resolution value other than 1. They just are not used

**Set Spline Resolution**
- Add Catmull Rom as one of the supported spline types
- Add a note that the resolution will not have an effect on segments between vector handles

**Set Spline Type**
- Add a note about Bezier handles after the conversion. (I think this seems a bit out of place here. Maybe add as a note?)
- Add Catmull Rom as one of the available types

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/104473