Makefile: add new  'make checkout_locale'
This adds a python script to do a sparse checkout of the manual translations.
We want to use a sparse checkout because checking out the whole translation directory can take a long time to download.
This will help save bandwidth and also makes it simpler for translators who do not need access to every language

For now just the script is committed, the contributing guides will be updated later after more user testing.