Add more information to the "Bake Curve" operator
It was never quite clear what the use case for the  "Bake Curve" operator was.
I looked at the code recently and the main use case seems to be to save disk space.
Additionally I left a note that it is a destructive process,
which many users are confused about as well.

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/104533