Crop Nodes Tools image
This is not really needed, but we want to investigate a bug.

We know the images won't show up on the PR, since we don't transfer the
LFS data to the main repo (as we do for the commits).

What we don't know is what happens when we merged this via the web-interface.

Upstream bug: https://github.com/go-gitea/gitea/issues/17715

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/104590