Extensions: Initial documentation
This is an updated subset of !104698.

These two pages will be linked from the extensions website directly.
They are committed ahead of the other changes so people can read the
documentation in context. But also because it makes it easy to roll them
back if needed without all the other changes.

The images don't represent the current state of the user interface, but
instead some changes I would like to see there, so they may change
still.

Note: This commit deliberately introduces the following warnings:
* manual/editors/preferences/extensions.rst: WARNING: document isn't included in any toctree
* manual/extensions/index.rst: WARNING: document isn't included in any toctree