Extensions: Add-ons page
Section about legacy add-ons, and how to convert them to extensions.

Note, I wanted to do:

However RST doesn't allow me to use anchor to another page in the doc (I
think). Suggestions welcome.

I consider doing an absolute link to the page, but that doesn't work
well if you are navigating in a non-english language, or not in /dev/.