Fix local language names in language menu
For Finnish, "Suomeksi" appears to be a declension of "Suomi", whose
nominative form is more commonly used to refer to the language.

For Ukrainian, it should be written in Cyrillic and use the word for
the language instead of the country. Issue reported by Marina
Veselkova.

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/104720