Update 3D Print Toolbox documentation
I've recently contributed a new feature to the `3D Print Toolbox` add-on and I wanted to add some  documentation of it in the Blender manual.

Then I realized the add-on manual was somewhat outdated, so I gave a shot at updating the entire entry.

Summary of the changes:
 - Add description of Hollow tool.
 - Add description of Align XY tool.
 - Small text corrections.
 - Correct reference to `3dprint_distorted.jpg` image. The target image was removed in a previous commit, probably because it wasn't used anywhere, so I restored it.

This is my first time contributing to the manual, and I am not a native English speaker, so a review is a good idea.

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/104761