Restore Rigify add-on documentation
Rigify was removed as part of the migration to extensions.blender.org,
then added in as core add-on again, so this change brings back its
documentation.

In addition, CloudRig is no longer a Rigify feature set, and it's not
mentioned in the user preferences, so it should not be here either.

This reverts commit 0df6e1f17cc7064304cae3ead2d624bca26e166f.

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/104852