Extensions: Clarified license page and update its requirements
The original implementation of the platform was too permissive on the accepted licenses.

The new recommendation from the Blender Foundation is to only support a
single license on the extensions platform for add-ons.

Since Blender itself is distributed under GNU/GPL 3.0, this is the
direction we are going for the platform as well. This means add-ons that
have Apache 2.0 code inside are compatible.

Ref: !104898