Anim: Updates to the Action Editor page
This patch updates the page for the action editor with the recent changes
* The Up/Down arrows have been removed, so the description of them
was removed as well. Left a note to trace that.
This also removes the images used there.
* Stash and Push down have been moved into the Action menu.
So their manual entry was moved as well
* Added description for the operators that work
with slots (merge, separate, extract)

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/105087