Fix: Update Action diagrams to work with dark theme
The approach taken is just to add a translucent background.

Additionally, switching to SVG instead of PNG for the images. The
original source SVG's were lost, which made this update annoying, and
there isn't really much reason not to just use straight SVG images
anyway. This way they can be more easily updated in the future.