UI: Add Blender version label for display
Currently, the variable `blender_version` is used for display in
titles and text, as well as for building paths and links.

In the case of LTS versions, the display should mention it, while the
paths/links should keep using the version number.

This PR adds the variable `blender_version_label`, which most of the
time will be the version number, except for LTS releases. And uses
it in the page title, index, and introduces the `BLENDER_VERSION_LABEL`
shorthand to be able to use it in text.

|main|PR|
|----|----|
|![main](/attachments/f1bc056b-f65b-47f2-8eae-7ab02a28518d)|![PR](/attachments/a1493c44-cfb6-45d6-97ce-2e4a0a70e196)|

Part of the efforts to improve Blender LTS branding blender/blender#124511

Co-authored-by: Aaron Carlisle <blendify@noreply.localhost>
Pull Request: https://projects.blender.org/blender/blender-manual/pulls/105233