Paint: Change usages of 'Radius' to 'Size'
This commit specifically changes mentions of the 'Radius' property to
'Size' where applicable. Notably, it does not change areas where we
refer to the brush radius as a concept, such as when we talk about
vertices within the brush radius or ratios related to it.

Pull Request: https://projects.blender.org/blender/blender-manual/pulls/105315