Sharp edges are labeled as redundant with smooth shading
As far as I see it Sharp Edges are not redundant when an object is shaded as smooth - because they stay sharp even with smooth shading.
For flat shading they are redundant, but not for smooth shading.

Please correct me if I have a wrong understanding of the matter.

Co-authored-by: Aaron Carlisle <blendify@noreply.localhost>
Pull Request: https://projects.blender.org/blender/blender-manual/pulls/105288